Three bugs a sanitizer and a duplicate key found (v0.1.160)
Went looking with tools rather than by reading. The compiler at -Wall -Wextra
plus a dozen extra warnings had nothing to say, so: UBSan and ASan over the
render paths, a randomised agreement check on the model, and degenerate inputs
(empty timeline, zero-length clip, 1x1 frame, blur radius bigger than the image)
through every effect and every transition.

**Two commands sharing a key silently killed both.** Qt calls two QShortcuts with
the same key on one window ambiguous and fires NEITHER. The panel refuses to
create that — it asks before reassigning — but nothing else did: importing a
profile, or a settings file left behind by a build where a command was renamed,
lands the same key on two commands and both features go dead while the panel
cheerfully shows them as bound. rebuild() now gives a key to at most one command,
and bindings arriving from a file are de-duplicated on the way in so what the
panel shows is what actually works. Registration order decides, which is also the
order the panel lists them in.

harpia/tests/shortcut_dupkey_test.cpp drives the real registry with real key
events. On the old code it reports A=0 B=0 — both dead — and on the new one
exactly one command runs and the other stops claiming a key it cannot use.

**Signed integer overflow in the Dissolve transition.** hash01() multiplied a
signed int by 374761393, which overflows from x = 6 onwards, so every pixel of
every dissolve past the seventh column was undefined behaviour. It happens to
wrap the way the code wants on the compilers used today; it is not required to,
and the optimiser is allowed to assume it never happens. Unsigned now — the same
fix the identical hash in EffectClip.cpp got in v0.1.155, which this one was
copied from before that. UBSan flagged it and is clean after.

**std::clamp with the bounds the wrong way round.** Timeline::setStart clamps to
`end_ - 1`, which is -1 before setDuration() has run, and setEnd clamps up from
`start_ + 1`, which can exceed duration_. clamp with hi < lo is undefined, not
merely wrong. Latent — no current caller reaches it — but these are public
setters and the guard is one std::max each.

Everything else came back clean, which is worth recording too: ASan and UBSan
over the snap, multi-track snap, gutter, transition, spotlight and effect-clip
tests found nothing; 287,018 randomised comparisons confirmed the O(n log n)
overlap sweep added in v0.1.155 agrees with the original per-clip version on
every arrangement, including equal start times and nested clips; and no effect
or transition crashes on a 1x1 frame, a null image or a mismatched pair.

harpia/tests/run_units.sh runs the fast checks that need no media.

+++ harpia/editor/timeline/Transitions.cpp
@@ -22,9 +22,13 @@ const char *kNames[kTransitionTypeCount] = {
 
 // Deterministic per-pixel threshold for the noise dissolve, so two exports of
 // the same project match.
+// Unsigned throughout: these multiplies are MEANT to wrap, and wrapping a
+// signed int is undefined behaviour the compiler may assume never happens. It
+// overflows from x = 6 onwards, so every pixel of a dissolve past the seventh
+// column was relying on it.
 inline double hash01(int x, int y)
 {
-	unsigned int n = (unsigned int)(x * 374761393 + y * 668265263);
+	unsigned int n = unsigned(x) * 374761393u + unsigned(y) * 668265263u;
 	n = (n ^ (n >> 13)) * 1274126177u;
 	return double((n ^ (n >> 16)) & 0xffff) / 65535.0;
 }
